Cleaning up code, changing file names, etc
