--- utilities/Colors.js
@@ -1,16 +1,0 @@
-const Colors = {
-    white: '#fff',
-    black: '#000',
-  
-    darkgrey: '#6D7365',
-    grey: '#9A9A9A',
-    lightgrey: '#E0E0E0',
-  
-    green: '#648C1C',
-    lightgreen: '#CAF28D',
-    mutedgreen: '#effbdd',
-  
-    lightblue: '#82A2F4'
-  }
-  
-  export default Colors
